requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
filelock==3.20.1
freetype-py==2.5.1
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyglet==2.1.12
PyOpenGL==3.1.0
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
trimesh==4.10.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wheel==0.45.1
